requirements.txt
streamlit==1.29.0
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
xgboost==3.0.0
pillow>=10.0.0